reuse filter and getTestCases
